labhouse/requirements.txt
accelerate==0.26.1
alembic==1.13.1
amqp==5.2.0
anyio==4.2.0
asgiref==3.7.2
asttokens==2.4.1
async-timeout==4.0.3
Authlib==1.2.1
azure-common==1.1.28
azure-core==1.29.6
azure-identity==1.15.0
azure-mgmt-core==1.4.0
azure-mgmt-rdbms==10.1.0
azure-mgmt-resource==23.0.1
azure-mgmt-subscription==3.1.1
Babel==2.14.0
bcrypt==4.0.1
bidict==0.22.1
billiard==4.2.0
blinker==1.7.0
boto3==1.33.13
botocore==1.33.13
Brotli==1.1.0
cachetools==5.3.2
celery==5.3.6
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cryptography==41.0.7
decorator==5.1.1
diffusers==0.26.1
Django==4.1.13
django-cors-headers==4.3.1
djangorestframework==3.14.0
dnspython==2.4.2
email-validator==2.1.0.post1
eventlet==0.34.2
exceptiongroup==1.2.0
executing==2.0.1
filelock==3.13.1
Flask==2.3.3
flask-babel==4.0.0
Flask-Compress==1.14
Flask-Gravatar==0.5.0
Flask-Login==0.6.3
Flask-Mail==0.9.1
Flask-Migrate==4.0.5
Flask-Paranoid==0.3.0
Flask-Principal==0.4.0
Flask-Security-Too==5.2.0
Flask-SocketIO==5.3.6
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
fsspec==2023.12.2
google-api-core==2.15.0
google-api-python-client==2.113.0
google-auth==2.26.2
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.1.0
googleapis-common-protos==1.62.0
greenlet==1.1.2
gunicorn==21.2.0
h11==0.14.0
httpagentparser==1.9.5
httplib2==0.22.0
huggingface-hub==0.20.3
idna==3.6
importlib-metadata==7.0.1
importlib-resources==6.1.1
ipdb==0.13.13
ipython==8.20.0
isodate==0.6.1
itsdangerous==2.1.2
jaraco.classes==3.3.0
jedi==0.19.1
jeepney==0.8.0
Jinja2==3.1.3
jmespath==1.0.1
keyring==24.3.0
kombu==5.3.4
ldap3==2.9.1
Mako==1.3.0
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
mdurl==0.1.2
more-itertools==10.2.0
mpmath==1.3.0
msal==1.26.0
msal-extensions==1.1.0
msrest==0.7.1
networkx==3.2.1
numpy==1.26.3
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
packaging==23.2
pandas==2.1.4
paramiko==3.4.0
parso==0.8.3
passlib==1.7.4
pexpect==4.9.0
pillow==10.2.0
portalocker==2.8.2
prompt-toolkit==3.0.43
protobuf==4.25.2
psutil==5.9.7
psycopg-binary==3.1.12
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
Pygments==2.17.2
PyJWT==2.8.0
PyNaCl==1.5.0
pyotp==2.9.0
pyparsing==3.1.1
pypng==0.20220715.0
pyscopg2==66.0.2
python-dateutil==2.8.2
python-engineio==4.8.2
python-socketio==5.11.0
pytz==2023.3.post1
PyYAML==6.0.1
qrcode==7.4.2
redis==5.0.1
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.7.0
rsa==4.9
s3transfer==0.8.2
safetensors==0.4.2
SecretStorage==3.3.3
shellingham==1.5.4
simple-websocket==1.0.0
six==1.16.0
sniffio==1.3.0
speaklater3==1.4
SQLAlchemy==2.0.25
sqlparse==0.2.4
sshtunnel==0.4.0
stack-data==0.6.3
sympy==1.12
tokenizers==0.15.1
tomli==2.0.1
torch==2.2.0
torchvision==0.17.0
tqdm==4.66.1
traitlets==5.14.1
transformers==4.37.2
triton==2.2.0
typer==0.9.0
typing_extensions==4.9.0
tzdata==2023.4
ua-parser==0.18.0
uritemplate==4.1.1
urllib3==1.26.18
user-agents==2.2.0
vine==5.1.0
wcwidth==0.2.13
Werkzeug==2.3.8
wsproto==1.2.0
WTForms==3.1.2
zipp==3.17.0
